Created GeneralService packages and moved the classes
