+++ src/main/java/GeneralService/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package GeneralService;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
